refactor: modularize code and added detailed comments
