+++ crates/core/generate-pie/src/lib.rs
@@ -61,306 +61,30 @@
 //! - **Output Options**: Configure output file paths and formats
 
 use cairo_vm::types::layout_name::LayoutName;
-use rpc_client::state_reader::AsyncRpcStateReader;
 use rpc_client::RpcClient;
 use starknet::core::types::BlockId;
-use starknet_api::core::{ChainId, ContractAddress};
-use starknet_os::io::os_output::StarknetOsRunnerOutput;
 use starknet_os::{
     io::os_input::{OsChainInfo, OsHints, OsHintsConfig, StarknetOsInput},
     runner::run_os_stateless,
 };
-use starknet_types_core::felt::Felt;
 use std::path::Path;
 
-/// Default timeout for RPC requests in seconds.
-pub const DEFAULT_RPC_TIMEOUT_SECONDS: u64 = 30;
-
-/// Maximum number of blocks that can be processed in a single PIE generation.
-pub const MAX_BLOCKS_PER_PIE: usize = 100;
-
-/// Default layout name for Cairo execution.
-pub const DEFAULT_CAIRO_LAYOUT: &str = "all_cairo";
-
-/// Default Sepolia STRK fee token address.
-pub const DEFAULT_SEPOLIA_STRK_FEE_TOKEN: &str = "0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d";
+use block_processor::collect_single_block_info;
+use cached_state::generate_cached_state_input;
+use error::PieGenerationError;
+use types::{PieGenerationInput, PieGenerationResult};
 
 mod api_to_blockifier_conversion;
 mod block_processor;
 mod cached_state;
 mod commitment_utils;
+mod constants;
 mod context_builder;
 mod error;
 mod rpc_utils;
 mod state_processing;
 mod state_update;
-
-use block_processor::collect_single_block_info;
-use cached_state::generate_cached_state_input;
-
-/// Configuration for chain-specific settings.
-///
-/// This struct contains all the chain-specific configuration needed for PIE generation,
-/// including the chain ID, fee token addresses, and whether the chain is an L3.
-///
-/// # Examples
-///
-/// ```rust
-/// use generate_pie::ChainConfig;
-/// use starknet_api::core::ChainId;
-///
-/// // Use default Sepolia configuration
-/// let config = ChainConfig::default();
-///
-/// // Create custom configuration
-/// let custom_config = ChainConfig {
-///     chain_id: ChainId::Mainnet,
-///     strk_fee_token_address: ContractAddress::try_from(Felt::from_hex_unchecked("0x123...")).unwrap(),
-///     is_l3: true,
-/// };
-/// ```
-#[derive(Debug, Clone, PartialEq)]
-pub struct ChainConfig {
-    /// The chain ID for the target Starknet network.
-    pub chain_id: ChainId,
-    /// The address of the STRK fee token contract.
-    pub strk_fee_token_address: ContractAddress,
-    /// Whether this is an L3 chain (true) or L2 chain (false).
-    pub is_l3: bool,
-}
-
-impl Default for ChainConfig {
-    /// Creates a default configuration for Sepolia testnet.
-    ///
-    /// # Returns
-    ///
-    /// A `ChainConfig` instance with Sepolia defaults:
-    /// - Chain ID: Sepolia
-    /// - STRK fee token: Sepolia STRK token address
-    /// - L3: false (L2 chain)
-    fn default() -> Self {
-        Self {
-            chain_id: ChainId::Sepolia,
-            strk_fee_token_address: ContractAddress::try_from(Felt::from_hex_unchecked(DEFAULT_SEPOLIA_STRK_FEE_TOKEN))
-                .expect("Valid Sepolia STRK fee token address"),
-            is_l3: false,
-        }
-    }
-}
-
-impl ChainConfig {
-    /// Validates the chain configuration.
-    ///
-    /// # Returns
-    ///
-    /// Returns `Ok(())` if the configuration is valid, or an error if validation fails.
-    ///
-    /// # Errors
-    ///
-    /// Returns a `PieGenerationError::InvalidConfig` if the configuration is invalid.
-    pub fn validate(&self) -> Result<(), PieGenerationError> {
-        // Validate that the STRK fee token address is not zero
-        if self.strk_fee_token_address == ContractAddress::default() {
-            return Err(PieGenerationError::InvalidConfig("STRK fee token address cannot be zero".to_string()));
-        }
-
-        Ok(())
-    }
-}
-
-/// Configuration for OS hints and execution parameters.
-///
-/// This struct controls various aspects of the Starknet OS execution, including
-/// debug mode, output verbosity, and data availability mode.
-///
-/// # Examples
-///
-/// ```rust
-/// use generate_pie::OsHintsConfiguration;
-///
-/// // Use default configuration
-/// let config = OsHintsConfiguration::default();
-///
-/// // Create custom configuration for debugging
-/// let debug_config = OsHintsConfiguration {
-///     debug_mode: true,
-///     full_output: true,
-///     use_kzg_da: false,
-/// };
-/// ```
-#[derive(Debug, Clone, PartialEq)]
-pub struct OsHintsConfiguration {
-    /// Whether to enable debug mode for detailed logging and output.
-    pub debug_mode: bool,
-    /// Whether to generate full output including intermediate states.
-    pub full_output: bool,
-    /// Whether to use KZG (Kate-Zaverucha-Goldberg) data availability mode.
-    pub use_kzg_da: bool,
-}
-
-impl Default for OsHintsConfiguration {
-    /// Creates a default configuration with sensible defaults.
-    ///
-    /// # Returns
-    ///
-    /// A `OsHintsConfiguration` instance with:
-    /// - Debug mode: enabled (for better error reporting)
-    /// - Full output: disabled (for performance)
-    /// - KZG DA: enabled (modern data availability)
-    fn default() -> Self {
-        Self { debug_mode: true, full_output: false, use_kzg_da: true }
-    }
-}
-
-impl OsHintsConfiguration {
-    /// Validates the OS hints configuration.
-    ///
-    /// # Returns
-    ///
-    /// Returns `Ok(())` if the configuration is valid, or an error if validation fails.
-    ///
-    /// # Errors
-    ///
-    /// Returns a `PieGenerationError::InvalidConfig` if the configuration is invalid.
-    pub fn validate(&self) -> Result<(), PieGenerationError> {
-        // Currently no validation needed, but this provides a place for future validation
-        Ok(())
-    }
-}
-
-/// Input configuration for PIE generation.
-///
-/// This struct contains all the necessary configuration and parameters needed to
-/// generate a Cairo PIE file from Starknet blocks.
-///
-/// # Examples
-///
-/// ```rust
-/// use generate_pie::{PieGenerationInput, ChainConfig, OsHintsConfiguration};
-///
-/// let input = PieGenerationInput {
-///     rpc_url: "https://your-starknet-node.com".to_string(),
-///     blocks: vec![12345, 12346],
-///     chain_config: ChainConfig::default(),
-///     os_hints_config: OsHintsConfiguration::default(),
-///     output_path: Some("output.pie".to_string()),
-/// };
-/// ```
-#[derive(Debug, Clone)]
-pub struct PieGenerationInput {
-    /// The RPC URL of the Starknet node to connect to.
-    pub rpc_url: String,
-    /// The list of block numbers to process for PIE generation.
-    pub blocks: Vec<u64>,
-    /// Chain-specific configuration settings.
-    pub chain_config: ChainConfig,
-    /// OS hints and execution configuration.
-    pub os_hints_config: OsHintsConfiguration,
-    /// Optional output file path for the generated PIE file.
-    pub output_path: Option<String>,
-}
-
-impl PieGenerationInput {
-    /// Validates the PIE generation input configuration.
-    ///
-    /// # Returns
-    ///
-    /// Returns `Ok(())` if the input is valid, or an error if validation fails.
-    ///
-    /// # Errors
-    ///
-    /// Returns a `PieGenerationError::InvalidConfig` if the input is invalid.
-    pub fn validate(&self) -> Result<(), PieGenerationError> {
-        // Validate RPC URL
-        if self.rpc_url.trim().is_empty() {
-            return Err(PieGenerationError::InvalidConfig("RPC URL cannot be empty".to_string()));
-        }
-
-        // Validate blocks
-        if self.blocks.is_empty() {
-            return Err(PieGenerationError::InvalidConfig("At least one block must be specified".to_string()));
-        }
-
-        // Validate maximum number of blocks
-        if self.blocks.len() > MAX_BLOCKS_PER_PIE {
-            return Err(PieGenerationError::InvalidConfig(format!(
-                "Too many blocks specified: {} (maximum: {})",
-                self.blocks.len(),
-                MAX_BLOCKS_PER_PIE
-            )));
-        }
-
-        // Validate that blocks are in ascending order
-        let mut sorted_blocks = self.blocks.clone();
-        sorted_blocks.sort();
-        if sorted_blocks != self.blocks {
-            return Err(PieGenerationError::InvalidConfig("Blocks must be specified in ascending order".to_string()));
-        }
-
-        // Validate chain configuration
-        self.chain_config.validate()?;
-
-        // Validate OS hints configuration
-        self.os_hints_config.validate()?;
-
-        Ok(())
-    }
-}
-
-/// Result containing the generated PIE and metadata.
-///
-/// This struct contains the output of the PIE generation process, including
-/// the generated Cairo PIE, information about processed blocks, and the output path.
-pub struct PieGenerationResult {
-    /// The output from the Starknet OS execution containing the Cairo PIE.
-    pub output: StarknetOsRunnerOutput,
-    /// The list of block numbers that were successfully processed.
-    pub blocks_processed: Vec<u64>,
-    /// The output file path where the PIE was saved (if specified).
-    pub output_path: Option<String>,
-}
-
-/// Main error type for PIE generation.
-///
-/// This enum represents all possible errors that can occur during the PIE generation
-/// process, including block processing errors, RPC client errors, OS execution errors,
-/// and configuration errors.
-#[derive(thiserror::Error, Debug)]
-pub enum PieGenerationError {
-    /// Block processing failed for a specific block.
-    #[error("Block processing failed for block {block_number}: {source}")]
-    BlockProcessing {
-        /// The block number that failed to process.
-        block_number: u64,
-        /// The underlying error that caused the failure.
-        #[source]
-        source: Box<dyn std::error::Error + Send + Sync>,
-    },
-
-    /// RPC client-related error.
-    #[error("RPC client error: {0}")]
-    RpcClient(String),
-
-    /// OS execution related error.
-    #[error("OS execution error: {0}")]
-    OsExecution(String),
-
-    /// I/O error during file operations.
-    #[error("IO error: {0}")]
-    Io(#[from] std::io::Error),
-
-    /// Invalid configuration error.
-    #[error("Invalid configuration: {0}")]
-    InvalidConfig(String),
-
-    /// State processing error.
-    #[error("State processing error: {0}")]
-    StateProcessing(String),
-
-    /// Contract class processing error.
-    #[error("Contract class processing error: {0}")]
-    ContractClassProcessing(String),
-}
+pub mod types;
 
 /// Core function to generate PIE from blocks.
 ///
@@ -426,10 +150,6 @@ pub async fn generate_pie(input: PieGenerationInput) -> Result<PieGenerationResu
     // Process each block
     for (index, block_number) in input.blocks.iter().enumerate() {
         log::info!("=== Processing block {} ({}/{}) ===", block_number, index + 1, input.blocks.len());
-
-        // Create a state reader for the current block
-        let _blockifier_state_reader = AsyncRpcStateReader::new(rpc_client.clone(), BlockId::Number(*block_number));
-        log::debug!("State reader created for block {}", block_number);
 
         // Collect block information
         log::info!("Starting to collect block info for block {}", block_number);
